refactor: 메서드 명 및 메서드 위치 변경
