+++ src/main/java/roomescape/controller/ReservationController.java
@@ -29,7 +29,7 @@ public class ReservationController {
 
     @GetMapping
     public ResponseEntity<List<ReservationResponse>> getReservations() {
-        List<Reservation> reservations = reservationService.allReservations();
+        List<Reservation> reservations = reservationService.findAll();
         List<ReservationResponse> responses = reservations.stream()
                 .map(ReservationResponse::from)
                 .toList();
