+++ src/main/java/roomescape/service/ThemeService.java
@@ -1,8 +1,11 @@
 package roomescape.service;
 
+import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.domain.Reservation;
 import roomescape.domain.Theme;
+import roomescape.domain.ThemeRanking;
 import roomescape.dto.request.CreateThemeRequest;
 import roomescape.exception.InvalidThemeException;
 import roomescape.repository.ReservationRepository;
@@ -10,6 +13,9 @@ import roomescape.repository.ThemeRepository;
 
 @Service
 public class ThemeService {
+
+    private static final int THEME_RANKING_END_RANGE = 7;
+    private static final int THEME_RANKING_START_RANGE = 1;
 
     private final ThemeRepository themeRepository;
     private final ReservationRepository reservationRepository;
@@ -38,5 +44,14 @@ public class ThemeService {
             throw new InvalidThemeException("예약이 존재하는 테마는 삭제할 수 없습니다.");
         }
         themeRepository.deleteById(id);
+    }
+
+    public List<Theme> getRankingThemes(LocalDate originDate) {
+        LocalDate end = originDate.minusDays(THEME_RANKING_START_RANGE);
+        LocalDate start = end.minusDays(THEME_RANKING_END_RANGE);
+        List<Reservation> inRangeReservations = reservationRepository.findAllByDateInRange(start, end);
+
+        ThemeRanking themeRanking = new ThemeRanking(inRangeReservations);
+        return themeRanking.getAscendingRanking();
     }
 }
