+++ src/test/java/roomescape/unit/service/ThemeServiceTest.java
@@ -2,9 +2,11 @@ package roomescape.unit.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import roomescape.domain.LoginMember;
@@ -35,12 +37,13 @@ class ThemeServiceTest {
     private ReservationTimeService reservationTimeService;
 
     private ThemeRepository themeRepository;
+    private ReservationTimeRepository reservationTimeRepository;
     private MemberRepository memberRepository;
 
     @BeforeEach
     void setUp() {
-        ReservationTimeRepository reservationTimeRepository = new FakeReservationTimeRepository();
         ReservationRepository reservationRepository = new FakeReservationRepository();
+        reservationTimeRepository = new FakeReservationTimeRepository();
         themeRepository = new FakeThemeRepository();
         memberRepository = new FakeMemberRepository();
 
@@ -109,5 +112,54 @@ class ThemeServiceTest {
     void 존재하지_않는_테마를_조회시_예외가_발생한다() {
         assertThatThrownBy(() -> themeService.getThemeById(1L))
                 .isInstanceOf(InvalidThemeException.class);
+    }
+
+    @Test
+    void 최근_일주일을_기준으로_예약이_많은_테마_10개를_확인할_수_있다() {
+        // given
+        final int THEME_COUNT = 10;
+        final int TIME_SLOTS = 6;
+
+        for (int i = 0; i < THEME_COUNT; i++) {
+            Theme theme = new Theme( "테마" + (i + 1), "테마", "thumbnail");
+            themeRepository.add(theme);
+        }
+
+        for (int i = 0; i < TIME_SLOTS; i++) {
+            LocalTime localTime = LocalTime.of(10 + i, 0);
+            reservationTimeRepository.add(new ReservationTime( localTime));
+        }
+
+        Member beforeAddMember = new Member( "Hula", "test@test.com", "test", Role.USER);
+        Member member = memberRepository.add(beforeAddMember);
+        LoginMember loginMember = new LoginMember(member.getId(), member.getName(), member.getRole());
+
+        // 테마 1 예약 3개
+        reservationService.addReservation(new CreateReservationRequest(LocalDate.now().plusDays(1), 1L, 1L),
+                loginMember);
+        reservationService.addReservation(new CreateReservationRequest(LocalDate.now().plusDays(1), 2L, 1L),
+                loginMember);
+        reservationService.addReservation(new CreateReservationRequest(LocalDate.now().plusDays(1), 3L, 1L),
+                loginMember);
+
+        // 테마 2 예약 2개
+        reservationService.addReservation(new CreateReservationRequest(LocalDate.now().plusDays(1), 1L, 2L),
+                loginMember);
+        reservationService.addReservation(new CreateReservationRequest(LocalDate.now().plusDays(1), 2L, 2L),
+                loginMember);
+
+        // 테마 3 예약 1개
+        reservationService.addReservation(new CreateReservationRequest(LocalDate.now().plusDays(1), 1L, 3L),
+                loginMember);
+
+        // when
+        List<Theme> themeRanking = themeService.getRankingThemes(LocalDate.now().plusDays(6));
+
+        // then
+        assertAll(() -> {
+            assertThat(themeRanking.getFirst().getId()).isEqualTo(1L);
+            assertThat(themeRanking.get(1).getId()).isEqualTo(2L);
+            assertThat(themeRanking.get(2).getId()).isEqualTo(3L);
+        });
     }
 }
